+++ docker-compose.dev.yml
@@ -2,7 +2,7 @@ version: '3.9'
 
 services:
   ipinfo:
-    image: ghcr.io/skidoodle/ipinfo:main
+    build: .
     container_name: ipinfo
     restart: unless-stopped
     ports:
